backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
pymysql==1.1.0
cryptography==42.0.5
python-dotenv==1.0.0
openai>=1.0.0
pydantic>=2.0.0
sqlalchemy>=2.0.0
